move crypto to objects rename package to customhostname (follow convention)
