--- src/main/java/eu/roboflax/cloudflare/crypto/customHostname/Method.java
@@ -1,6 +1,0 @@
-package eu.roboflax.cloudflare.objects.crypto.customHostname;
-
-public enum Method {
-
-  http, email, cname
-}
