--- src/main/java/eu/roboflax/cloudflare/crypto/customHostname/Status.java
@@ -1,5 +1,0 @@
-package eu.roboflax.cloudflare.objects.crypto.customHostname;
-
-public enum Status {
-  on, off
-}
