--- src/main/java/eu/roboflax/cloudflare/crypto/customHostname/Type.java
@@ -1,5 +1,0 @@
-package eu.roboflax.cloudflare.objects.crypto.customHostname;
-
-public enum Type {
-  dv // domain validation
-}
