+++ src/main/java/eu/roboflax/cloudflare/objects/crypto/customhostname/Method.java
@@ -1,0 +1,6 @@
+package eu.roboflax.cloudflare.objects.crypto.customhostname;
+
+public enum Method {
+
+  http, email, cname
+}
