+++ src/main/java/eu/roboflax/cloudflare/objects/crypto/customhostname/Ssl.java
@@ -1,4 +1,4 @@
-package eu.roboflax.cloudflare.objects.crypto.customHostname;
+package eu.roboflax.cloudflare.objects.crypto.customhostname;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
