+++ src/main/java/eu/roboflax/cloudflare/objects/crypto/customhostname/SslStatus.java
@@ -1,4 +1,4 @@
-package eu.roboflax.cloudflare.objects.crypto.customHostname;
+package eu.roboflax.cloudflare.objects.crypto.customhostname;
 
 public enum SslStatus {
   initializing, pending_validation, pending_issuance, pending_deployment, active
