+++ src/main/java/eu/roboflax/cloudflare/objects/crypto/customhostname/Status.java
@@ -1,0 +1,5 @@
+package eu.roboflax.cloudflare.objects.crypto.customhostname;
+
+public enum Status {
+  on, off
+}
